Add wallet iframe view
