+++ layouts/assets.vue
@@ -1,0 +1,5 @@
+<template>
+  <div class="min-h-screen px-28 py-20">
+    <slot />
+  </div>
+</template>
