+++ pages/[profileAddress]/assets.vue
@@ -1,0 +1,9 @@
+<script setup lang="ts">
+definePageMeta({
+  layout: 'assets',
+})
+</script>
+
+<template>
+  <Assets />
+</template>
